Support browsers argument
A sequence of labels specifying the browsers to include.

Usually, any dependency that you wish to be included in the runfiles
tree should be included within the data attribute. However, data
dependencies, by default, are copied to the Bazel output tree before
being passed as inputs to runfiles.

This is not a good default behavior for browser since these typically
come from external workspaces which cannot be symlinked into bazel-bin.
Instead, we place them at the root of the runfiles tree. Use relative
paths to construct account for this placement